+++ CHANGES.md
@@ -1,3 +1,6 @@
+## next
+ * bugfix angular.module("MyMod").bootstrap(e, [], {}) disrupting chaining
+
 ## v0.8.0 2014-05-09
  * ngRoute support: $routeProvider.when("path", { .. })
  * even faster (~11% faster annotating angular.js)